Update vite to 4.5.0 and move config to ./build/vite
