+++ package.json
@@ -26,8 +26,9 @@
   },
   "devDependencies": {
     "pnpm": "8.7.5",
-    "zx": "7.2.3",
+    "vite": "4.5.0",
     "vite-node": "0.34.6",
+    "zx": "7.2.3",
     "prettier": "3.0.3",
     "pretty-quick": "3.1.3",
     "clang-format": "1.8.0"
