+++ sites/dev/package.json
@@ -16,7 +16,7 @@
     "ol": "^8",
     "coi-serviceworker": "0.1.7",
     "typescript": "5.0.4",
-    "vite": "^4.4.9",
+    "vite": "4.5.0",
     "vite-plugin-static-copy": "0.17.0"
   }
 }
