+++ sites/dev/vite.config.ts
@@ -1,9 +1,9 @@
 import { resolve } from "path";
 import { defineConfig } from "vite";
 
-import QgisRuntimePlugin from "../../vite/QgisRuntimePlugin";
-import CrossOriginIsolationPlugin from "../../vite/CrossOriginIsolationPlugin";
-import DirectoryListingPlugin from "../../vite/DirectoryListingPlugin";
+import QgisRuntimePlugin from "../../build/vite/QgisRuntimePlugin";
+import CrossOriginIsolationPlugin from "../../build/vite/CrossOriginIsolationPlugin";
+import DirectoryListingPlugin from "../../build/vite/DirectoryListingPlugin";
 
 import { viteStaticCopy } from "vite-plugin-static-copy";
